clean

+++ src/GenerateCSR.php
@@ -102,7 +102,7 @@ EOL;
 
         openssl_csr_export($csr, $csrAsString);
 
-        if (isset($this->opensslConfig['config'])) {
+        if (isset($this->opensslConfig['config']) && file_exists($this->opensslConfig['config'])) {
             unlink($this->opensslConfig['config']);
         }
